requirements.txt
# Core Data Science Libraries (Python 3.12 compatible)
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Machine Learning Models
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Time Series Analysis (optional - can be installed later if needed)
# prophet==1.1.4  # May have compatibility issues with Python 3.12
# statsmodels>=0.14.0

# Recommender Systems
scikit-surprise>=1.1.3

# Model Interpretability
shap>=0.43.0
# lime==0.2.0.1  # May have issues, can install later if needed

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Database
mysql-connector-python>=8.1.0
SQLAlchemy>=2.0.20
pymysql>=1.1.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.1
joblib>=1.3.2
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Development Tools
black>=23.10.0
flake8>=6.1.0
ipython>=8.17.0
jupyter>=1.0.0

# Production
gunicorn>=21.2.0
